Stricter adherence to Apple developer guidelines!

+++ lib/TH/generic/THTensorRandom.c
@@ -17,7 +17,7 @@ TH_API void THTensor_(random)(THTensor *self)
 #elif defined(TH_REAL_IS_FLOAT)
   TH_TENSOR_APPLY(real, self, *self_data = (float)(THRandom_random() % ((1UL << FLT_MANT_DIG)+1)););
 #elif defined(TH_REAL_IS_DOUBLE)
-  TH_TENSOR_APPLY(real, self, *self_data = (float)(THRandom_random() % ((1UL << DBL_MANT_DIG)+1)););
+  TH_TENSOR_APPLY(real, self, *self_data = (double)(THRandom_random() % ((1ULL << DBL_MANT_DIG)+1)););
 #else
 #error "Unknown type"
 #endif
